agentic-tool-ifc-parsing-martinasimoni-dotcom/requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
ifcopenshell>=0.7.0
pytest>=7.0.0
